requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-cloud-texttospeech==2.14.1
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
